Factor out common code from the three new tests
