+++ src/test/java/org/osc/controller/nsc/AbstractSampleSdnPluginTest.java
@@ -1,0 +1,46 @@
+/*******************************************************************************
+ * Copyright (c) Intel Corporation
+ * Copyright (c) 2017
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *******************************************************************************/
+package org.osc.controller.nsc;
+
+import static org.osc.controller.nsc.TestData.setupDataObjects;
+
+import javax.persistence.EntityManager;
+
+import org.junit.After;
+import org.junit.Before;
+import org.mockito.Answers;
+import org.mockito.Mock;
+
+public class AbstractSampleSdnPluginTest {
+
+    protected EntityManager em;
+
+    @Mock(answer = Answers.CALLS_REAL_METHODS)
+    protected TestTransactionControl txControl;
+
+    @Before
+    public void setup() {
+        this.em = InMemDB.getEntityManager();
+        this.txControl.init(this.em);
+        setupDataObjects();
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        InMemDB.close();
+    }
+}
